requirements.txt
streamlit
mag4
pandas
matplotlib.pyplot